dbus-vis: fix comparison of different signedness
Some places in the code compare signed and unsigned integers, for
example:

  if (choice_ >= items_.size())

This generates a compiler warning "warning: comparison of integer
expressions of different signedness: 'int' and
'std::vector<std::__cxx11::basic_string<char> >::size_type' {aka
'unsigned int'} [-Wsign-compare]".

This fix adds a static_cast<int> to the size() types to fix the
warnings.

+++ dbus-top/menu.cpp
@@ -71,7 +71,7 @@ void ArrowKeyNavigationMenu::do_Render(bool is_column_major)
                 wattrset(win_, A_REVERSE);
             }
             std::string s = items_[idx];
-            while (s.size() < col_width_)
+            while (static_cast<int>(s.size()) < col_width_)
             {
                 s.push_back(' ');
             }
@@ -299,7 +299,7 @@ void ArrowKeyNavigationMenu::AddItem(const std::string& s)
 
 bool ArrowKeyNavigationMenu::RemoveHighlightedItem(std::string* ret)
 {
-    if (choice_ < 0 || choice_ >= items_.size())
+    if (choice_ < 0 || choice_ >= static_cast<int>(items_.size()))
         return false;
     std::string r = items_[choice_];
     items_.erase(items_.begin() + choice_);
@@ -309,7 +309,7 @@ bool ArrowKeyNavigationMenu::RemoveHighlightedItem(std::string* ret)
     }
     else
     {
-        if (choice_ >= items_.size())
+        if (choice_ >= static_cast<int>(items_.size()))
         {
             choice_ = items_.size() - 1;
         }
